+++ src/MWSMerchantFulfillment.php
@@ -15,7 +15,6 @@ class MWSMerchantFulfillment
 
     public function getEligibleShippingServices($params = [])
     {
-        $nextToken = data_get($params, 'NextToken');
         $action = 'GetEligibleShippingServices';
 
         $response = $this->client->post($action, '/MerchantFulfillment/'.self::VERSION, self::VERSION, $params);
@@ -34,6 +33,53 @@ class MWSMerchantFulfillment
         ];
 
         return $data;
+    }
+
+    public function createShipment($params = [])
+    {
+        $action = 'CreateShipment';
+
+        $response = $this->client->post($action, '/MerchantFulfillment/'.self::VERSION, self::VERSION, $params);
+
+        return $this->parseResponse($response, $action.'Result');
+    }
+
+    public function cancelShipment($shipmentId)
+    {
+        $action = 'CancelShipment';
+
+        $response = $this->client->post($action, '/MerchantFulfillment/'.self::VERSION, self::VERSION, [
+            'ShipmentId' => $shipmentId
+        ]);
+
+        return $this->parseResponse($response, $action.'Result');
+    }
+
+    public function getShipment($shipmentId)
+    {
+        $action = 'GetShipment';
+
+        $response = $this->client->post($action, '/MerchantFulfillment/'.self::VERSION, self::VERSION, [
+            'ShipmentId' => $shipmentId
+        ]);
+
+        return $this->parseResponse($response, $action.'Result');
+    }
+
+    public function getServiceStatus()
+    {
+        $action = 'GetServiceStatus';
+
+        $response = $this->client->post($action, '/MerchantFulfillment/'.self::VERSION, self::VERSION);
+        return $this->parseResponse($response, $action.'Result');
+    }
+
+    public function getAdditionalSellerInputs($params)
+    {
+        $action = 'GetAdditionalSellerInputs';
+
+        $response = $this->client->post($action, '/MerchantFulfillment/'.self::VERSION, self::VERSION, $params);
+        return $this->parseResponse($response, $action.'Result');
     }
 
 }