+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add Merchant Fulfillment Service Class
 - Add Get Eligible Shipping Services Method
+- Add Get Shipment Method
+- Add Create Shipment Method
+- Add Cancel Shipment Method
+- Add Get Additional Seller Inputs Method
+- Add Get Service Status Method (Merchant Fulfillment)
+- Add readme notice for Amazon Seller Rest Api
 ## [0.2.0] - 2020-10-09
 ### Updated
 - Add Laravel 8 Support